Fix hooks
