+++ packages/graph-explorer/src/hooks/index.ts
@@ -2,5 +2,9 @@ export { default as useDebounceValue } from "./useDebounceValue";
 export { default as useDeepMemo } from "./useDeepMemo";
 export { default as useEntities } from "./useEntities";
 export { default as useExpandNode } from "./useExpandNode";
-export { default as useFetchNode } from "./useFetchNode";
+export * from "./useAddToGraph";
+export * from "./useRemoveFromGraph";
+export { default as useDisplayNames } from "./useDisplayNames";
+export { default as useTextTransform } from "./useTextTransform";
 export { default as usePrevious } from "./usePrevious";
+export { default as useTranslations } from "./useTranslations";
