+++ packages/graph-explorer/src/modules/EntitiesTabular/components/NodesTabular.tsx
@@ -19,10 +19,12 @@ import {
   nodesSelectedIdsAtom,
 } from "@/core/StateProvider/nodes";
 
-import { useDeepMemo } from "@/hooks";
-import useDisplayNames from "@/hooks/useDisplayNames";
-import useTextTransform from "@/hooks/useTextTransform";
-import useTranslations from "@/hooks/useTranslations";
+import {
+  useDeepMemo,
+  useDisplayNames,
+  useTextTransform,
+  useTranslations,
+} from "@/hooks";
 import { recoilDiffSets } from "@/utils/recoilState";
 
 type ToggleVertex = Vertex & { __is_visible: boolean };
