+++ packages/graph-explorer/src/modules/EntityDetails/EdgeDetail.tsx
@@ -20,8 +20,7 @@ import EdgeIcon from "@/components/icons/EdgeIcon";
 import { useWithTheme } from "@/core";
 import { useConfiguration } from "@/core/ConfigurationProvider";
 import fade from "@/core/ThemeProvider/utils/fade";
-import useDisplayNames from "@/hooks/useDisplayNames";
-import useTextTransform from "@/hooks/useTextTransform";
+import { useTextTransform, useDisplayNames } from "@/hooks";
 import { formatDate } from "@/utils";
 import defaultStyles from "./EntityDetail.styles";
 import { useEdgeTypeConfig } from "@/core/ConfigurationProvider/useConfiguration";
