+++ packages/graph-explorer/src/modules/GraphViewer/internalComponents/ContextMenu.tsx
@@ -17,9 +17,7 @@ import { useWithTheme } from "@/core";
 import { edgesSelectedIdsAtom, toEdgeMap } from "@/core/StateProvider/edges";
 import { nodesSelectedIdsAtom, toNodeMap } from "@/core/StateProvider/nodes";
 import { userLayoutAtom } from "@/core/StateProvider/userPreferences";
-import useDisplayNames from "@/hooks/useDisplayNames";
-import useEntities from "@/hooks/useEntities";
-import useTranslations from "@/hooks/useTranslations";
+import { useDisplayNames, useEntities, useTranslations } from "@/hooks";
 import useGraphGlobalActions from "../useGraphGlobalActions";
 import defaultStyles from "./ContextMenu.styles";
 import { EdgeId, VertexId } from "@/@types/entities";
